Repository and Implementation
