+++ app/src/main/java/com/example/crearrepositorio/domain/SeriesRepository.kt
@@ -1,0 +1,7 @@
+package com.example.crearrepositorio.domain
+
+import kotlinx.coroutines.flow.Flow
+
+interface SeriesRepository {
+    fun getAllSeries(): Flow<List<SerieModel>>
+    }
